Refactor extraction planner to improve Collection support
This was a major overhaul of the extraction planner that was designed
to improve support for Initializable Collections. The majority of the work
involved consisted of moving knowledge of what it means to be a Collection
into a new kind of `ExtractionPlanNode`, namely a `CollectionInitializerNode`.

RegExtract previously had a fairly unnatural notion of List<T>, blending
both the List-ness of a node with the underlying type T. Even weirder,
ListOfListNode was a whole other animal, that handled nesting of containers.

There is now better support for list patterns of either of the following
forms:
1. @"((\d+),? ?)+", or
2. @"(?:(\d+),? ?)+"

Previously, only the latter type of pattern (with a non-capturing group) was
supported if you needed to select out a substring from within the pattern
repeated inside the list.